Moving to DATACLAY/2.4.dev
This commit cleans up a lot of things, prepares DATACLAY/2.4
(currently DATACLAY/2.4.dev, official release will come soon)
and has tracing ready.

Things that this includes:

 - register.sh script is no longer used, as `dataclayprepare`
 sript simplifies things

 - no need to change the `.bashrc`: DATACLAY module behaves
 correctly with the current loaded modules / environment.

 - same with PYTHON module: no need to load it explicitly

 - cleaned up some superfluous / incorrect `--classpath` and
 `--pythonpath` arguments on the `enqueue_compss` call

 - `STORAGE_PROPS` is used through a temporal file; this allows
 consecutive launches with different configurations (the file
 is read when task is started, not submitted).

 - necessary changes to activate simultaneous tracing in
 dataClay and COMPSs. This is achieved through a little bash
 scripting in the `launch_with_dataClay.sh` script. The
 rationale is that tracing flag should be a single activating
 flag and affect both dataClay and COMPSs.

It is a long list of things and this sounds as a
monster-improper-commit, but most things are very intertwined
and related to how dataClay 2.4 is being shaped. So this is,
in fact, a mostly atomic change from dataClay 2.0 to 2.4.